DDCNL-2249 | Move to build.sbt format and update dependecies and plugins
