--- project/Microservice.scala
@@ -1,85 +1,0 @@
-import play.sbt.routes.RoutesKeys.{routesGenerator, _}
-import sbt.Keys._
-import sbt.Tests.{Group, SubProcess}
-import sbt._
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin
-import uk.gov.hmrc.versioning.SbtGitVersioning
-import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
-import com.lucidchart.sbt.scalafmt.ScalafmtCorePlugin.autoImport.scalafmtOnCompile
-import play.sbt.PlayImport.PlayKeys
-
-trait MicroService {
-
-  import uk.gov.hmrc._
-  import DefaultBuildSettings._
-  import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin._
-  import uk.gov.hmrc.SbtAutoBuildPlugin
-  import uk.gov.hmrc.SbtArtifactory
-
-  val appName: String
-
-  lazy val appDependencies : Seq[ModuleID] = ???
-  lazy val playSettings : Seq[Setting[_]] = Seq.empty
-
-  lazy val scoverageSettings = {
-    import scoverage.ScoverageKeys
-    val scoverageExcludePatterns = List(
-      "<empty>",
-      "Reverse.*",
-      "app.Routes.*",
-      "uk.gov.hmrc.BuildInfo.*",
-      "prod.*",
-      "dev.*",
-      "uk.gov.hmrc.tai.config")
-
-    Seq(
-      ScoverageKeys.coverageExcludedPackages := scoverageExcludePatterns.mkString("", ";", ""),
-      ScoverageKeys.coverageMinimum := 97,
-      ScoverageKeys.coverageFailOnMinimum := true,
-      ScoverageKeys.coverageHighlighting := true,
-      parallelExecution in Test := false
-    )
-  }
-
-  lazy val microservice = Project(appName, file("."))
-    .enablePlugins(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory)
-    .settings(playSettings ++ scoverageSettings : _*)
-    .settings(publishingSettings: _*)
-    .settings(
-      libraryDependencies ++= appDependencies,
-      retrieveManaged := true,
-      evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false),
-      routesGenerator := InjectedRoutesGenerator,
-      scalafmtOnCompile := true,
-      PlayKeys.playDefaultPort := 9331
-    )
-    .settings(inConfig(ITTestPhases.TemplateTest)(Defaults.testSettings): _*)
-    .configs(IntegrationTest)
-    .settings(inConfig(ITTestPhases.TemplateItTest)(Defaults.itSettings): _*)
-    .settings(
-      Keys.fork in IntegrationTest := false,
-      unmanagedSourceDirectories in IntegrationTest <<= (baseDirectory in IntegrationTest)(base => Seq(base / "it")),
-      testGrouping in IntegrationTest := ITTestPhases.oneForkedJvmPerTest((definedTests in IntegrationTest).value),
-      parallelExecution in IntegrationTest := false)
-    .settings(
-      resolvers += Resolver.jcenterRepo,
-      scalacOptions ++= Seq("-unchecked", "-deprecation", "-feature", "-Xlint"),
-      routesImport ++= Seq("scala.language.reflectiveCalls", "uk.gov.hmrc.tai.model.domain.income._", "uk.gov.hmrc.tai.model.domain._")
-    )
-    .settings(majorVersion := 0)
-
-}
-
-private object ITTestPhases {
-
-  val allPhases = "tt->test;test->test;test->compile;compile->compile"
-  val allItPhases = "tit->it;it->it;it->compile;compile->compile"
-
-  lazy val TemplateTest = config("tt") extend Test
-  lazy val TemplateItTest = config("tit") extend IntegrationTest
-
-  def oneForkedJvmPerTest(tests: Seq[TestDefinition]) =
-    tests map {
-      test => new Group(test.name, Seq(test), SubProcess(ForkOptions(runJVMOptions = Seq("-Dtest.name=" + test.name))))
-    }
-}
